# src/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Healthcare Management System API
  version: 1.0.0
  description: |
    HIPAA-compliant Healthcare Management System with comprehensive patient, appointment, and medical record management.
    
    ## Features
    - Patient management with full demographic information
    - Appointment scheduling and management
    - Medical record creation and retrieval with encryption
    - Patient consent management for HIPAA compliance
    - Secure authentication using JWT tokens
    - Comprehensive audit logging
    
    ## Security
    All endpoints require JWT Bearer token authentication.
    
    ## Compliance
    This API is designed to meet HIPAA compliance requirements including:
    - Data encryption at rest and in transit
    - Comprehensive audit trails
    - Access control and authorization
    - Patient consent management
  contact:
    name: Healthcare System Support
    email: support@healthcare.com
    url: https://healthcare.com
  license:
    name: Proprietary
    url: https://healthcare.com/license
servers:
  - url: http://localhost:8080
    description: Development Server
  - url: https://api.healthcare.com
    description: Production Server
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/appointments:
    post:
      tags:
        - Appointment Management
      summary: Schedule a new appointment
      description: Creates a new appointment for a patient with a healthcare provider
      operationId: scheduleAppointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentDTO'
      responses:
        '201':
          description: Appointment scheduled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentDTO'
        '400':
          description: Invalid input data or time slot unavailable
        '401':
          description: Unauthorized
        '404':
          description: Patient not found
  /api/v1/appointments/{id}:
    get:
      tags:
        - Appointment Management
      summary: Get appointment by ID
      description: Retrieves appointment details by appointment ID
      operationId: getAppointmentById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Appointment ID
      responses:
        '200':
          description: Appointment found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentDTO'
        '401':
          description: Unauthorized
        '404':
          description: Appointment not found
...
  /api/v1/consents:
    post:
      tags:
        - Consent Management
      summary: Record a new patient consent
      description: Records a new consent from a patient for treatment, disclosure, research, or marketing purposes
      operationId: recordConsent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsentDTO'
      responses:
        '201':
          description: Consent recorded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentDTO'
        '400':
          description: Invalid input data
        '401':
          description: Unauthorized
        '404':
          description: Patient not found
  /api/v1/consents/{id}:
    get:
      tags:
        - Consent Management
      summary: Get consent by ID
      description: Retrieves a specific consent record by its unique identifier
      operationId: getConsentById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Consent ID
      responses:
        '200':
          description: Consent found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentDTO'
        '401':
          description: Unauthorized
        '404':
          description: Consent not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  schemas:
...
    AppointmentDTO:
      type: object
      required:
        - patientId
        - providerId
        - providerName
        - appointmentDateTime
        - durationMinutes
        - appointmentType
        - status
      properties:
        id:
          type: integer
          format: int64
          description: Appointment ID
          example: 1
        patientId:
          type: integer
          format: int64
          description: Patient ID
          example: 123
        providerId:
          type: string
          description: Healthcare provider identifier
          example: PROV123
        providerName:
          type: string
          description: Healthcare provider name
          example: Dr. Jane Smith
        appointmentDateTime:
          type: string
          format: date-time
          description: Appointment date and time
          example: '2024-02-15T14:30:00'
        durationMinutes:
          type: integer
          description: Appointment duration in minutes
          example: 30
        appointmentType:
          type: string
          enum: [CONSULTATION, FOLLOW_UP, EMERGENCY, ROUTINE, SPECIALIST]
          description: Type of appointment
          example: CONSULTATION
        status:
          type: string
          enum: [SCHEDULED, CONFIRMED, CANCELLED, COMPLETED, NO_SHOW]
          description: Appointment status
          example: SCHEDULED
        reason:
          type: string
          description: Reason for appointment
        notes:
          type: string
          description: Additional notes
...
    ConsentDTO:
      type: object
      required:
        - patientId
        - consentType
        - granted
      properties:
        id:
          type: integer
          format: int64
          description: Consent ID
          example: 1
        patientId:
          type: integer
          format: int64
          description: Patient ID
          example: 123
        consentType:
          type: string
          enum: [TREATMENT, DISCLOSURE, RESEARCH, MARKETING]
          description: Type of consent
          example: TREATMENT
        granted:
          type: boolean
          description: Consent granted status
          example: true
        description:
          type: string
          description: Consent description or terms
        expirationDate:
          type: string
          format: date-time
          description: Timestamp when consent expires
          example: '2025-01-15T10:30:00'
        obtainedBy:
          type: string
          description: Person who obtained the consent
          example: Dr. Smith